image-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-vision
PyPDF2
python-magic
requests
pydantic
